WIP: task_manager -> task_runner_impl, TaskManager -> TaskRunner
Eventually (maybe sooner than we thought!) the code that was/is in
task_manager.nim, threadpool.nim, etc. (the fully generic modules) will become
part of the nim-task-runner repo. The renaming done in this commit is intended
to help make that outlook more apparent.

Note that `tasks/qt.nim` is an example of specialization so it would remain in
the status-desktop repo.